Fix issue where `.size` property is called as function
`GXByteBuffer.size` is a property with getters and setters, calling it as function will
throw exception as it calls the value returned from the getter (in this case an int).

This fixes it only in one place, there are 27 to a function named  with over half of them
seemingly used on a GXByteBuffer object.